Renamed linter lib
It makes no functional difference, but it lets the `package.json` and the require in the specs look more similar between all the packages.
