+++ lib/init.js
