+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "linter-gotype",
   "linter-package": true,
-  "main": "./lib/linter-gotype",
+  "main": "./lib/init",
   "version": "0.1.0",
   "private": true,
   "description": "Check Go files with gotype tool",
